Divide a base form and schema for future use
