+++ app/forms/support_form_schema.rb
@@ -1,6 +1,6 @@
-require "dry/validation"
+# require "dry/validation"
 
-Dry::Validation.load_extensions(:predicates_as_macros)
+# Dry::Validation.load_extensions(:predicates_as_macros)
 
 #
 # Validate "support request" form details from a "school business professional"
@@ -9,13 +9,7 @@ Dry::Validation.load_extensions(:predicates_as_macros)
 #
 # @author Peter Hamilton
 #
-class SupportFormSchema < Dry::Validation::Contract
-  import_predicates_as_macros
-
-  config.messages.backend = :i18n
-  config.messages.top_namespace = :forms
-  config.messages.load_paths << Rails.root.join("config/locales/validation/en.yml")
-
+class SupportFormSchema < Schema
   params do
     optional(:phone_number).value(:string)  # step 1
     optional(:journey_id).value(:string)    # step 2
